Adición de enpoints de busqueda por ID y actualizado

+++ src/main/java/pe/edu/upc/inmobivalor_be/repositories/IUsuarioRepository.java
@@ -9,7 +9,7 @@ import org.springframework.stereotype.Repository;
 import pe.edu.upc.inmobivalor_be.entities.Usuario;
 
 @Repository
-public interface IUsuarioRepository extends JpaRepository<Usuario, Long> {
+public interface IUsuarioRepository extends JpaRepository<Usuario, Integer> {
     public Usuario findOneByUsername(String username);
